+++ R/data_desc.r
@@ -138,3 +138,104 @@
 #' data(graded)
 #' summary(graded)
 "graded"
+
+#' example0.rda
+#'
+#' A generated data set for rasch FLPS to mimic CTA1 data.
+#'
+#' @format A data frame with variables:
+#' \describe{
+#'   \item{Y}{Outcome}
+#'   \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
+#'   \item{sex}{0 = boys; 1 = girls}
+#'   \item{race}{0 = White; 1 = Others}
+#'   \item{pretest}{Pre test scores}
+#'   \item{stdscore}{Standardized scores}
+#'   \item{V1}{Binary item}
+#'   \item{V2}{Binary item}
+#'   \item{V3}{Binary item}
+#'   \item{V4}{Binary item}
+#'   \item{V5}{Binary item}
+#'   \item{V6}{Binary item}
+#'   \item{V7}{Binary item}
+#'   \item{V8}{Binary item}
+#'   \item{V9}{Binary item}
+#'   \item{V10}{Binary item}
+#'   \item{V11}{Binary item}
+#'   \item{V12}{Binary item}
+#' }
+"example0"
+
+
+#' example1.rda
+#'
+#' A generated data set containing binary items information.
+#'
+#' @format A data frame with variables:
+#' \describe{
+#'   \item{Y}{Outcome}
+#'   \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
+#'   \item{X1}{Continuous covariate 1}
+#'   \item{X2}{Continuous covariate 2}
+#'   \item{q1}{Binary item}
+#'   \item{q2}{Binary item}
+#'   \item{q3}{Binary item}
+#'   \item{q4}{Binary item}
+#'   \item{q5}{Binary item}
+#'   \item{q6}{Binary item}
+#'   \item{q7}{Binary item}
+#'   \item{q8}{Binary item}
+#'   \item{q9}{Binary item}
+#'   \item{q10}{Binary item}
+#' }
+"example1"
+
+#' example2.rda
+#'
+#' A generated data set containing continuous items information.
+#'
+#' @format A data frame with variables:
+#' \describe{
+#'   \item{Y}{Outcome}
+#'   \item{trt}{Treatment assignment; 0 = control, 1 = treatment}
+#'   \item{X1}{Continuous covariate 1}
+#'   \item{X2}{Continuous covariate 2}
+#'   \item{q1}{Continuous item}
+#'   \item{q2}{Continuous item}
+#'   \item{q3}{Continuous item}
+#'   \item{q4}{Continuous item}
+#'   \item{q5}{Continuous item}
+#'   \item{q6}{Continuous item}
+#'   \item{q7}{Continuous item}
+#'   \item{q8}{Continuous item}
+#'   \item{q9}{Continuous item}
+#'   \item{q10}{Continuous item}
+#' }
+"example2"
+
+#' example3.rda
+#'
+#' A generated data set for multidimensional FLPS.
+#'
+#' @format A data frame with variables:
+#' \describe{
+#'   \item{Y}{Outcome}
+#'   \item{Z}{Treatment assignment; 0 = control, 1 = treatment}
+#'   \item{X1}{Continuous covariate 1}
+#'   \item{X2}{Continuous covariate 2}
+#'   \item{X3}{Continuous covariate 2}
+#'   \item{X4}{Continuous covariate 2}
+#'   \item{V1}{Continuous item}
+#'   \item{V2}{Continuous item}
+#'   \item{V3}{Continuous item}
+#'   \item{V4}{Continuous item}
+#'   \item{V5}{Continuous item}
+#'   \item{V6}{Continuous item}
+#'   \item{V7}{Continuous item}
+#'   \item{V8}{Continuous item}
+#'   \item{V9}{Continuous item}
+#'   \item{V10}{Continuous item}
+#'   \item{V11}{Continuous item}
+#'   \item{V12}{Continuous item}
+#' }
+"example3"